refactor: decouple the workflow stepper cluster from app viewmodelbuilders and unions (#1617)
