+++ app/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders.tsx
@@ -1,6 +1,4 @@
 import { SLUGIFY_OPTIONS } from "@/common/constants";
-import { type StepConfig } from "@/components/Entity/components/ConfigureWorkflowInputs/components/Main/components/Stepper/components/Step/types";
-import { type ConfiguredInput } from "@/views/WorkflowInputsView/hooks/UseConfigureInputs/types";
 import { type BRCDataCatalogGenome } from "@brc/apis/assembly";
 import { type BRCDataCatalogOrganism } from "@brc/apis/organism";
 import { type Outbreak } from "@brc/apis/outbreak";
@@ -12,15 +10,11 @@ import { type ResourcesSection } from "@brc/views/PriorityPathogenView/component
 import { type KeyValueSection } from "@brc/views/PriorityPathogenView/ui/Section/KeyValueSection/keyValueSection";
 import { type MDXSection } from "@brc/views/PriorityPathogenView/ui/Section/MDXSection/mdxSection";
 import { type Breadcrumb } from "@databiosphere/findable-ui/lib/components/common/Breadcrumbs/breadcrumbs";
-import { KeyElType } from "@databiosphere/findable-ui/lib/components/common/KeyValuePairs/components/KeyElType/keyElType";
-import { ValueElType } from "@databiosphere/findable-ui/lib/components/common/KeyValuePairs/components/ValueElType/valueElType";
 import {
   type Key,
   type KeyValuePairs,
   type Value,
 } from "@databiosphere/findable-ui/lib/components/common/KeyValuePairs/keyValuePairs";
-import { Stack } from "@databiosphere/findable-ui/lib/components/common/Stack/stack";
-import { TypographyWordBreak } from "@databiosphere/findable-ui/lib/components/common/Typography/TypographyWordBreak/TypographyWordBreak";
 import { type BackPageHero } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/components/BackPageHero/backPageHero";
 import { type Link } from "@databiosphere/findable-ui/lib/components/Links/components/Link/link";
 import { COLUMN_IDENTIFIER } from "@databiosphere/findable-ui/lib/components/Table/common/columnIdentifier";
@@ -33,7 +27,6 @@ import type {
   AssemblyContract,
   OrganismContract,
 } from "@repo/shared/apis/types";
-import type { Workflow } from "@repo/shared/apis/workflow";
 import { AppLink } from "@repo/shared/components/AppLink/appLink";
 import { AnalyzeGenome } from "@repo/shared/components/Table/components/TableCell/components/AnalyzeGenome/analyzeGenome";
 import { LevelCell } from "@repo/shared/components/Table/components/TableCell/components/LevelCell/levelCell";
@@ -1016,57 +1009,6 @@ function buildOrganismGenomesTableColumns(): ColumnDef<BRCDataCatalogGenome>[] {
     },
   ];
 }
-
-/**
- * Build props for the workflow configuration KeyValuePairs component.
- * @param configuredInput - Configured inputs.
- * @param configuredSteps - Configured steps.
- * @returns Props to be used for the KeyValuePairs component.
- */
-export const buildWorkflowConfiguration = (
-  configuredInput: ConfiguredInput,
-  configuredSteps: StepConfig[]
-): ComponentProps<typeof KeyValuePairs> => {
-  const keyValuePairs = new Map<Key, Value>();
-  for (const key of Object.keys(configuredInput)) {
-    // Find the step config, for the configured input.
-    const stepConfig = configuredSteps.find((step) => step.key === key);
-    if (!stepConfig) continue;
-    // Get the value for the configured input.
-    const value = stepConfig.renderValue?.(configuredInput);
-    if (value === undefined) continue;
-    keyValuePairs.set(stepConfig.label, value);
-  }
-  // If there are no configured inputs, add a "None" value.
-  if (keyValuePairs.size === 0) {
-    keyValuePairs.set("", "None");
-  }
-  return {
-    KeyElType: KeyElType,
-    KeyValuesElType: (props) => <Stack {...props} gap={4} />,
-    ValueElType: TypographyWordBreak,
-    keyValuePairs,
-  };
-};
-
-/**
- * Build props for the workflow details KeyValuePairs component.
- * @param workflow - Workflow.
- * @returns Props to be used for the KeyValuePairs component.
- */
-export const buildWorkflowDetails = (
-  workflow: Workflow
-): ComponentProps<typeof KeyValuePairs> => {
-  const keyValuePairs = new Map<Key, Value>();
-  keyValuePairs.set("Workflow", workflow.workflowName);
-  keyValuePairs.set("Description", workflow.workflowDescription);
-  return {
-    KeyElType: KeyElType,
-    KeyValuesElType: (props) => <Stack {...props} gap={4} />,
-    ValueElType: ValueElType,
-    keyValuePairs,
-  };
-};
 
 /**
  * Returns an entity list link with a priority pathogen filter.
